// src/service/member/schemas/apiResponseEducationalClassResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Member API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { EducationalClassResult } from './educationalClassResult';

export interface ApiResponseEducationalClassResult {
  status: number;
  result?: EducationalClassResult;
  timestamp?: string;
  error?: string;
  debug?: string;
}